feat: Support launching pre-hook prior to launching services.
The pre-hook is treated identical to a service for the most part by all
the modules except that it runs always in the single-service mode, and
the final cleanup is just exiting all the goroutines in the janitor.

Create a new repo and janitor to manage the pre-hook, and after the
pre-hook completes, create a separate repo and janitor to manage the
services launched.

The signal manager switches to the new repo/janitor when moving from
pre-hook to managing the services.